requirements.txt
Flask==3.0.0
yfinance==1.1.0
APScheduler==3.10.4
python-dotenv==1.0.0
gspread==6.2.1
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0
